pyproject.toml
[tool.poetry]
name = "test-task-hq"
version = "0.1.0"
description = ""
authors = ["efsnail007 <sgefsnkiospch908@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
black = "^23.9.1"
isort = "^5.12.0"
pylint = "^2.17.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
